requirements.txt
accelerate==0.31.0
einops==0.7.0
fvcore==0.1.5.post20221221
h5py==3.11.0
librosa==0.10.2.post1
numpy==1.26.2
pandas==1.5.3
PyYAML==6.0.1
scikit_learn==1.3.2
scipy==1.11.4
simplejson==3.19.2
timm==0.4.12
tqdm==4.66.1
transformers==4.35.2
wget==3.2
cmapy
nlpaug
cmake==3.18.4
audiomentations
seaborn
torchaudio
torchcodec